requirements.txt
requests==2.19.1
prompt_toolkit==3.0.2
tableformatter==0.1.4
Keras==2.3.1
pytest==4.6.3
joblib==0.13.2
tabulate==0.8.3
gdown==3.8.3
python_nubia==0.2b2
allennlp==0.9.0
matplotlib==2.2.3
gensim==3.7.3
seaborn==0.10.0
Jinja2==2.10.3
termcolor==1.1.0
pandas==0.25.3
bert_serving_client==1.10.0
Pygments==2.2.0
natsort==6.2.0
numpy==1.18.1
pdfkit==0.6.1
scipy==1.2.1
bert_serving==0.0.1
GPUtil==1.4.0
scikit-learn==0.20.3
tensorflow==1.15.2
bokeh==1.0.4
nltk==3.4.5
umap-learn==0.3.10
